requirements.txt
# This template is a low-dependency template. 
# By default there is no requirements added here.
# Add the requirements you need to this file.
# or run `make init` to create this file automatically based on the template.
# You can also run `make switch-to-poetry` to use the poetry package manager.

i18nice==0.15.5
pyTelegramBotAPI==4.21.0
#git+https://github.com/hiddify-com/pyTelegramBotAPI.git@async_register_next_step_handler#egg=pyTelegramBotAPI
aiohttp
asyncio
watchdog
psutil
pillow

qrcode
requests
python-dotenv
asyncssh
paramiko
